Moved ContactResource model classes to businesslogic folder, to comply with the project structure strategy
